fix the timing of destroying the Gallery Database

+++ app/src/main/kotlin/com/simplemobiletools/gallery/pro/activities/MainActivity.kt
@@ -239,7 +239,10 @@ class MainActivity : SimpleActivity(), DirectoryOperationsListener {
             config.tempSkipDeleteConfirmation = false
             mTempShowHiddenHandler.removeCallbacksAndMessages(null)
             removeTempFolder()
-            GalleryDatabase.destroyInstance()
+
+            if (!config.showAll) {
+                GalleryDatabase.destroyInstance()
+            }
         }
     }
